+++ packages/plugin-button/web/src/createButtonData.ts
@@ -2,7 +2,11 @@ import { merge } from 'lodash';
 import { DEFAULT_CONFIG } from './constants';
 import { Node_Type } from 'wix-rich-content-common';
 import { convertNodeDataToDraft } from 'ricos-content/libs/toDraftData';
-import type { CreatePluginsDataMap, LINK_BUTTON_TYPE, ACTION_BUTTON_TYPE } from 'wix-rich-content-common';
+import type {
+  CreatePluginsDataMap,
+  LINK_BUTTON_TYPE,
+  ACTION_BUTTON_TYPE,
+} from 'wix-rich-content-common';
 
 export const createButtonData:
   | CreatePluginsDataMap[typeof LINK_BUTTON_TYPE]